// dashboard/src/utils/hasura-api/generated/schemas/createObjectRelationshipArgs.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing database operations, migrations, and metadata permissions
 * OpenAPI spec version: 1.0.0
 */
import type { CreateObjectRelationshipArgsUsing } from './createObjectRelationshipArgsUsing';
import type { TableIdentifier } from './tableIdentifier';

export interface CreateObjectRelationshipArgs {
  table: TableIdentifier;
  /** Name of the relationship */
  name: string;
  using: CreateObjectRelationshipArgsUsing;
  /** Comment for the relationship */
  comment?: string;
  /** Name of the source database */
  source: string;
}